Use dynamic-cache in materializer-general; cleanup lru_cache leftovers
