+++ crates/data-router/Cargo.toml
@@ -25,7 +25,6 @@ tracing_utils           = { path = "../utils/crates/tracing" }
 # Crates.io
 anyhow         = "1.0.40"
 async-trait    = "0.1.50"
-lru-cache      = "0.1.2"
 serde          = { version = "1.0.126", features = ["derive"] }
 serde_json     = "1.0.64"
 tokio          = { version = "1.6.1", features = ["macros"] }
