+++ crates/materializer-general/Cargo.toml
@@ -13,6 +13,7 @@ cdl_dto     = { path = "../dto" }
 misc_utils  = { path = "../utils/crates/misc" }
 rpc         = { path = "../rpc" }
 utils       = { path = "../utils" }
+cache       = { path = "../utils/crates/cache" }
 communication_utils     = { path = "../utils/crates/communication" }
 settings_utils          = { path = "../utils/crates/settings" }
 metrics_utils           = { path = "../utils/crates/metrics" }
@@ -23,7 +24,6 @@ anyhow      = "1.0.40"
 async-trait = "0.1.50"
 futures     = "0.3.15"
 itertools   = "0.10.0"
-lru-cache   = "0.1.2"
 serde       = { version = "1.0.126", features = ["derive"] }
 serde_json  = "1.0.64"
 tokio       = { version = "1.6.1", features = ["rt-multi-thread", "macros", "sync"] }
