+++ crates/query-router/Cargo.toml
@@ -22,7 +22,6 @@ tracing_utils           = { path = "../utils/crates/tracing", features = ["http"
 # Crates.io
 anyhow      = "1.0.40"
 async-trait = "0.1.50"
-lru-cache   = "0.1.2"
 serde       = { version = "1.0.126", features = ["derive"] }
 serde_json  = "1.0.64"
 uuid        = { version = "0.8.2", features = ["v4", "serde"] }
