+++ crates/query-router/src/main.rs
@@ -5,7 +5,7 @@ use warp::Filter;
 
 use cache::DynamicCache;
 use metrics_utils as metrics;
-use schema::SchemaMetadataFetcher;
+use schema::SchemaMetadataSupplier;
 use serde::Deserialize;
 use settings_utils::{load_settings, LogSettings, MonitoringSettings, RepositoryStaticRouting};
 use std::collections::HashMap;
@@ -52,7 +52,7 @@ async fn main() -> anyhow::Result<()> {
 
     let schema_registry_cache = Arc::new(Mutex::new(DynamicCache::new(
         settings.cache_capacity,
-        SchemaMetadataFetcher::boxed(settings.services.schema_registry_url),
+        SchemaMetadataSupplier::boxed(settings.services.schema_registry_url),
     )));
 
     let cache_filter = warp::any().map(move || schema_registry_cache.clone());
